backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
potrace==1.2.0
python-docx==0.8.11
reportlab==4.0.4
comtypes==1.2.0
pywin32==306
requests==2.31.0
